tests/wpt/tests/paint-timing/with-first-paint/first-contentful-bg-image.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP due to background image</title>
</head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="main"></div>
</body>

<footer>
<script>
setup({"hide_test_state": true});
async_test(function (t) {
    assert_implements(window.PerformancePaintTiming, "Paint Timing isn't supported.");
    const body = document.getElementsByTagName('body')[0];
    body.style.backgroundImage = 'url(../resources/circles.png)';
    window.onload = function() {
        function testPaintEntries() {
            const bufferedEntries = performance.getEntriesByType('paint');
            if (bufferedEntries.length < 2) {
                t.step_timeout(function() {
                    testPaintEntries();
                }, 20);
                return;
            }
            t.step(function() {
                assert_equals(bufferedEntries.length, 2, "There should be two paint timing instances.");
                assert_equals(bufferedEntries[0].entryType, "paint");
                assert_equals(bufferedEntries[0].name, "first-paint");
                assert_equals(bufferedEntries[1].entryType, "paint");
                assert_equals(bufferedEntries[1].name, "first-contentful-paint");
                t.done();
            });
        }
        t.step(function() {
            testPaintEntries();
        });
    };
}, "First contentful paint fires due to background image render.");
</script>
<footer>
</html>
